requirements.txt
deepface>=0.0.79
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.13.0
